Cleaned up code.  Renamed Micro files.
