+++ src/com/csci468/micro/Main.java
@@ -1,22 +1,16 @@
 package com.csci468.micro;
 
-import com.csci468.micro.scanner.MicroScanner;
-
 import java.io.IOException;
 
 public class Main {
 
     public static void main(String[] args) throws IOException {
-        String input = "";
-
-        if(args.length > 0)
-            input = args[0];
+        if(args.length > 0) {
+            String input = args[0];
+            Micro micro = new Micro(input);
+            micro.Scan();
+        }
         else
             System.out.println("You must supply an input.");
-
-        MicroScanner microScanner = new MicroScanner(input);
-        microScanner.Scan();
-
-
     }
 }
